Language.JavaScript.Parser

Documentation

parse

Arguments

:: String
The input stream (Javascript source code).

-> String
The name of the Javascript source (filename or input device).

-> Either String JSAST
An error or maybe the abstract syntax tree (AST) of zero or more Javascript statements, plus comments.

Parse one compound statement, or a sequence of simple statements. Generally used for interactive input, such as from the command line of an interpreter. Return comments in addition to the parsed statements.

readJs :: String -> JSAST

parseFile :: FilePath -> IO JSAST

Parse the given file. For UTF-8 support, make sure your locale is set such that "System.IO.localeEncoding" returns "utf8"

parseFileUtf8 :: FilePath -> IO JSAST

Parse the given file, explicitly setting the encoding to UTF8 when reading it

showStripped :: JSAST -> String

showStrippedMaybe :: Show a => Either a JSAST -> String

AST elements

data JSExpression

Constructors

JSIdentifier !JSAnnot !String
Terminals

JSDecimal !JSAnnot !String

JSLiteral !JSAnnot !String

JSHexInteger !JSAnnot !String

JSOctal !JSAnnot !String

JSStringLiteral !JSAnnot !String

JSRegEx !JSAnnot !String

JSArrayLiteral !JSAnnot ![JSArrayElement] !JSAnnot
lb, contents, rb

JSAssignExpression !JSExpression !JSAssignOp !JSExpression
lhs, assignop, rhs

JSCallExpression !JSExpression !JSAnnot !(JSCommaList JSExpression) !JSAnnot
expr, bl, args, rb

JSCallExpressionDot !JSExpression !JSAnnot !JSExpression
expr, dot, expr

JSCallExpressionSquare !JSExpression !JSAnnot !JSExpression !JSAnnot
expr, [, expr, ]

JSCommaExpression !JSExpression !JSAnnot !JSExpression
expression components

JSExpressionBinary !JSExpression !JSBinOp !JSExpression
lhs, op, rhs

JSExpressionParen !JSAnnot !JSExpression !JSAnnot
lb,expression,rb

JSExpressionPostfix !JSExpression !JSUnaryOp
expression, operator

JSExpressionTernary !JSExpression !JSAnnot !JSExpression !JSAnnot !JSExpression
cond, ?, trueval, :, falseval

JSFunctionExpression !JSAnnot !JSIdent !JSAnnot !(JSCommaList JSIdent) !JSAnnot !JSBlock
fn,name,lb, parameter list,rb,block`

JSMemberDot !JSExpression !JSAnnot !JSExpression
firstpart, dot, name

JSMemberExpression !JSExpression !JSAnnot !(JSCommaList JSExpression) !JSAnnot

JSMemberNew !JSAnnot !JSExpression !JSAnnot !(JSCommaList JSExpression) !JSAnnot
new, name, lb, args, rb

JSMemberSquare !JSExpression !JSAnnot !JSExpression !JSAnnot
firstpart, lb, expr, rb

JSNewExpression !JSAnnot !JSExpression
new, expr

JSObjectLiteral !JSAnnot !JSObjectPropertyList !JSAnnot
lbrace contents rbrace

JSUnaryExpression !JSUnaryOp !JSExpression

JSVarInitExpression !JSExpression !JSVarInitializer
identifier, initializer

data JSAnnot

Constructors

JSAnnot !TokenPosn ![CommentAnnotation]
Annotation: position and comment/whitespace information

JSAnnotSpace
A single space character

JSNoAnnot
No annotation

data JSBinOp

Constructors

JSBinOpAnd !JSAnnot

JSBinOpBitAnd !JSAnnot

JSBinOpBitOr !JSAnnot

JSBinOpBitXor !JSAnnot

JSBinOpDivide !JSAnnot

JSBinOpEq !JSAnnot

JSBinOpGe !JSAnnot

JSBinOpGt !JSAnnot

JSBinOpIn !JSAnnot

JSBinOpInstanceOf !JSAnnot

JSBinOpLe !JSAnnot

JSBinOpLsh !JSAnnot

JSBinOpLt !JSAnnot

JSBinOpMinus !JSAnnot

JSBinOpMod !JSAnnot

JSBinOpNeq !JSAnnot

JSBinOpOr !JSAnnot

JSBinOpPlus !JSAnnot

JSBinOpRsh !JSAnnot

JSBinOpStrictEq !JSAnnot

JSBinOpStrictNeq !JSAnnot

JSBinOpTimes !JSAnnot

JSBinOpUrsh !JSAnnot

data JSBlock

Constructors

JSBlock !JSAnnot ![JSStatement] !JSAnnot
lbrace, stmts, rbrace

data JSUnaryOp

Constructors

JSUnaryOpDecr !JSAnnot

JSUnaryOpDelete !JSAnnot

JSUnaryOpIncr !JSAnnot

JSUnaryOpMinus !JSAnnot

JSUnaryOpNot !JSAnnot

JSUnaryOpPlus !JSAnnot

JSUnaryOpTilde !JSAnnot

JSUnaryOpTypeof !JSAnnot

JSUnaryOpVoid !JSAnnot

data JSSemi

Constructors

JSSemi !JSAnnot

JSSemiAuto

data JSAssignOp

Constructors

JSAssign !JSAnnot

JSTimesAssign !JSAnnot

JSDivideAssign !JSAnnot

JSModAssign !JSAnnot

JSPlusAssign !JSAnnot

JSMinusAssign !JSAnnot

JSLshAssign !JSAnnot

JSRshAssign !JSAnnot

JSUrshAssign !JSAnnot

JSBwAndAssign !JSAnnot

JSBwXorAssign !JSAnnot

JSBwOrAssign !JSAnnot

data JSTryCatch

Constructors

JSCatch !JSAnnot !JSAnnot !JSExpression !JSAnnot !JSBlock
catch,lb,ident,rb,block

JSCatchIf !JSAnnot !JSAnnot !JSExpression !JSAnnot !JSExpression !JSAnnot !JSBlock
catch,lb,ident,if,expr,rb,block

data JSTryFinally

Constructors

JSFinally !JSAnnot !JSBlock
finally,block

JSNoFinally

data JSStatement

Constructors

JSStatementBlock !JSAnnot ![JSStatement] !JSAnnot !JSSemi
lbrace, stmts, rbrace, autosemi

JSBreak !JSAnnot !JSIdent !JSSemi
break,optional identifier, autosemi

JSConstant !JSAnnot !(JSCommaList JSExpression) !JSSemi
const, decl, autosemi

JSContinue !JSAnnot !JSIdent !JSSemi
continue, optional identifier,autosemi

JSDoWhile !JSAnnot !JSStatement !JSAnnot !JSAnnot !JSExpression !JSAnnot !JSSemi
do,stmt,while,lb,expr,rb,autosemi

JSFor !JSAnnot !JSAnnot !(JSCommaList JSExpression) !JSAnnot !(JSCommaList JSExpression) !JSAnnot !(JSCommaList JSExpression) !JSAnnot !JSStatement
for,lb,expr,semi,expr,semi,expr,rb.stmt

JSForIn !JSAnnot !JSAnnot !JSExpression !JSBinOp !JSExpression !JSAnnot !JSStatement
for,lb,expr,in,expr,rb,stmt

JSForVar !JSAnnot !JSAnnot !JSAnnot !(JSCommaList JSExpression) !JSAnnot !(JSCommaList JSExpression) !JSAnnot !(JSCommaList JSExpression) !JSAnnot !JSStatement
for,lb,var,vardecl,semi,expr,semi,expr,rb,stmt

JSForVarIn !JSAnnot !JSAnnot !JSAnnot !JSExpression !JSBinOp !JSExpression !JSAnnot !JSStatement
for,lb,var,vardecl,in,expr,rb,stmt

JSFunction !JSAnnot !JSIdent !JSAnnot !(JSCommaList JSIdent) !JSAnnot !JSBlock !JSSemi
fn,name, lb,parameter list,rb,block,autosemi

JSIf !JSAnnot !JSAnnot !JSExpression !JSAnnot !JSStatement
if,(,expr,),stmt

JSIfElse !JSAnnot !JSAnnot !JSExpression !JSAnnot !JSStatement !JSAnnot !JSStatement
if,(,expr,),stmt,else,rest

JSLabelled !JSIdent !JSAnnot !JSStatement
identifier,colon,stmt

JSEmptyStatement !JSAnnot

JSExpressionStatement !JSExpression !JSSemi

JSAssignStatement !JSExpression !JSAssignOp !JSExpression !JSSemi
lhs, assignop, rhs, autosemi

JSMethodCall !JSExpression !JSAnnot !(JSCommaList JSExpression) !JSAnnot !JSSemi

JSReturn !JSAnnot !(Maybe JSExpression) !JSSemi
optional expression,autosemi

JSSwitch !JSAnnot !JSAnnot !JSExpression !JSAnnot !JSAnnot ![JSSwitchParts] !JSAnnot !JSSemi
switch,lb,expr,rb,caseblock,autosemi

JSThrow !JSAnnot !JSExpression !JSSemi
throw val autosemi

JSTry !JSAnnot !JSBlock ![JSTryCatch] !JSTryFinally
try,block,catches,finally

JSVariable !JSAnnot !(JSCommaList JSExpression) !JSSemi
var|const, decl, autosemi

JSWhile !JSAnnot !JSAnnot !JSExpression !JSAnnot !JSStatement
while,lb,expr,rb,stmt

JSWith !JSAnnot !JSAnnot !JSExpression !JSAnnot !JSStatement !JSSemi
with,lb,expr,rb,stmt list

data JSSwitchParts

Constructors

JSCase !JSAnnot !JSExpression !JSAnnot ![JSStatement]
expr,colon,stmtlist

JSDefault !JSAnnot !JSAnnot ![JSStatement]
colon,stmtlist

data JSAST

Constructors

JSAstProgram ![JSStatement] !JSAnnot
source elements, tailing whitespace

JSAstStatement !JSStatement !JSAnnot

JSAstExpression !JSExpression !JSAnnot

JSAstLiteral !JSExpression !JSAnnot

data CommentAnnotation

Constructors

CommentA TokenPosn String

WhiteSpace TokenPosn String

NoComment

data TokenPosn

TokenPosn records the location of a token in the input text. It has three fields: the address (number of characters preceding the token), line number and column of a token within the file. Note: The lexer assumes the usual eight character tab stops.

Constructors

TokenPn !Int !Int !Int

tokenPosnEmpty :: TokenPosn

Pretty Printing

renderJS :: JSAST -> Builder

renderToString :: JSAST -> String

renderToText :: JSAST -> Text