Language.JavaScript.Parser.SrcLocation

Documentation

data TokenPosn

TokenPosn records the location of a token in the input text. It has three fields: the address (number of characters preceding the token), line number and column of a token within the file. Note: The lexer assumes the usual eight character tab stops.

Constructors

TokenPn !Int !Int !Int

tokenPosnEmpty :: TokenPosn